test: 테스트 클래스에서 설정 분리
